% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addspouses.R
\name{addspouses}
\alias{addspouses}
\title{Add Spouses}
\usage{
addspouses(df, spid = "spouse")
}
\arguments{
\item{df}{a data frame with at least a pidm column}

\item{spid}{if df is a local data frame, spid will be passed to the .id argument of bind_rows()}
}
\value{
a data frame with the same columns as df but with additional rows for any spouses. Spouse rows have identical rows to the original data. Union is used, so duplicates are removed.
}
\description{
Add Spouses
}
