% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.email.exclusions.R
\name{apply.email.exclusions}
\alias{apply.email.exclusions}
\title{Apply Email Exclusoins}
\usage{
apply.email.exclusions(
  df,
  idcol = "pidm",
  primary.only = F,
  include.students = F,
  sol = F,
  exclude.fall = F,
  exclude.spring = F
)
}
\arguments{
\item{df}{A data frame to be used as a mailing list.}

\item{idcol}{A character string describing the name of the id column. Defaults to "pidm". If prefemail, deceased, primdnrind and excl columns do not exist in df, they will be joined in from hallptbl using this column.}

\item{primary.only}{A logical value indicating wither only the primary contact in a couple be included. Defaults to TRUE.}

\item{include.students}{A logical value indicating whether or not students should be kept in the list. By default, students are excluded.}

\item{sol}{A logical vlaue indicating whether or not the mailing is a solicitation. Note that if sol is TRUE, a warning will be generated if both exclude.fall and exclude.spring are FALSE.}

\item{exclude.fall}{A logical value indicating whether or not constituents with a NFC, No Fall Contact, exclusion code should be excluded.}

\item{exclude.spring}{A logical value indicating whether or not constituents with a NSC, No Spring Contact, exclusion code should be excluded.}
}
\value{
a data frame filtered for a traditional mailing. If prefemail, deceased, primdnrind or excl don't already exist in df, they wil be added
}
\description{
Apply the standard Millikin email exclusions to a mailing list.
}
