% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.currency.R
\name{as.currency}
\alias{as.currency}
\title{Format as currency}
\usage{
as.currency(x)
}
\arguments{
\item{x}{a numeric vector to be formatted}
}
\value{
a vector of strings formated as currency, with commas and dollar signs.
}
\description{
Given a numeric vector, format it as currency, implicitly changing it to a string.
}
