% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.data.frame.named.vector.R
\name{as.data.frame.named.vector}
\alias{as.data.frame.named.vector}
\title{Coerce to a Data Frame}
\usage{
\method{as.data.frame}{named.vector}(x, valname = deparse(substitute(x)))
}
\arguments{
\item{x}{a named vector}

\item{valname}{a character string to be used as the new column name. Defaults to the name of x}
}
\value{
a two column data frame with nrow = length(x)
}
\description{
converts a named vector to a 2 column data frame with the names in the first column and values in the second columns. Throws an error if no names are found. Row names are stripped out of the final data frame.
}

