% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.percent.R
\name{as.percent}
\alias{as.percent}
\title{Format as percent}
\usage{
as.percent(x, digits = 2)
}
\arguments{
\item{x}{a numeric vector to be formatted}

\item{digits}{a single numerical value determining how many digits should be shown}
}
\value{
a vector of strings formated as percent, with commas and percent signs.
}
\description{
Given a numeric vector, format it as percent, implicitly changing it to a string.
}
