% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bannerize.pidm.R
\name{bannerize.pidm}
\alias{bannerize.pidm}
\title{Bannerize a Pidm}
\usage{
bannerize.pidm(x)
}
\arguments{
\item{x}{a vector to convert to Banner pidms}
}
\value{
a character vector where x is padded out to 8 characters with zeros
}
\description{
turn a vector into Banner pidms, by padding it out to 8 characters with zeros. Coerces numeric values to character.
}

