% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupstatus.R
\name{buildlongxref}
\alias{buildlongxref}
\title{buildlongxref}
\usage{
buildlongxref(..., calculateCH = T)
}
\arguments{
\item{...}{as many file locations as you have available. All csvs will be read in and bad rows dumped.}

\item{calculateCH}{a logical value determing whether or not CH relationships should be calculated from PA codes.}
}
\value{
a data frame with pidm, xref_pidm and xref_code columns in long format.
}
\description{
build a table of all available xrefs from existing csvs (called from XAXREF). Also pulls all spouse xrefs from hallp. Used by getrelationships().
}

