% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupstatus.R
\name{catadd}
\alias{catadd}
\title{CatAdd}
\usage{
catadd(addr = addr1, city = city, st = st, zip = zip, sep = " ")
}
\arguments{
\item{addr}{a referece to the address column}

\item{city}{a reference to the city column}

\item{st}{a reference to the state column}

\item{zip}{a reference to the zip code column}

\item{sep}{a string to be used to seperate the columns}
}
\value{
returns the 4 columns pasted together
}
\description{
Creates a concatenated address. Primarily for use in dplyr::mutate().
}
