% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.df.R
\name{clean.df}
\alias{clean.df}
\title{Clean the names of a data frame}
\usage{
clean.df(df, keep.underscore = TRUE)
}
\arguments{
\item{df}{a data frame}

\item{keep.underscore}{a logical value indicating whether underscores be kept or removed}
}
\value{
a data frame, or if dplyr is available, a tbl_df.
}
\description{
Clean up a data frame by moving all column names to lower case, removing spaces and punctuation. Also, when dplyr is available, return the data frame as a tbl_df.
}

