% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.na.R
\name{count.na}
\alias{count.na}
\title{Count.NA}
\usage{
count.na(x)
}
\arguments{
\item{an}{object to count. All bets are off if you don't pass a data frame or a vector.}
}
\value{
If x is a vector, returns a single numeric count of the NAs in the vector. If x is a data frame, returns a two-column data frame with counts for each column of the data frame.
}
\description{
Count the number of NA values in a data frame or vector.
}
