% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_in.R
\name{count_in}
\alias{count_in}
\title{Count in}
\usage{
count_in(x, y, by = NULL)
}
\arguments{
\item{x, }{y tbls to compare}

\item{by}{a character vector of variables to join by. If \code{NULL}, the default, \verb{*_join()} will do a natural join, using all variables with common names across the two tables}
}
\value{
a 3 row data frame with the number of ID variables that are in x but not in y, number in both, and number that are in y but not in x.
}
\description{
compare two objects to see how many rows are shared between them
}
