% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupstatus.R
\name{countif}
\alias{countif}
\title{Countif}
\usage{
countif(col)
}
\arguments{
\item{col}{a referece to the column to be counted.}
}
\value{
a integer vector of counts for each value in col.
}
\description{
returns a vector of counts for a vector. Uses table() to calculate counts. Primarily for use in dplyr::mutate().
}

