% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/matchfunctions.R
\name{createcounts}
\alias{createcounts}
\title{Create count columns}
\usage{
createcounts(x, y)
}
\arguments{
\item{x}{a data frame which needs pidms}

\item{y}{a data frame of reference data, usually hallp}
}
\value{
a data frame with nrows(x) with count columns appended
}
\description{
Creates COUNTIF style counts to determine whether 1:1 matching can be done on a particular field.
}

