% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createhallpwitholdnames.R
\name{createhallpwitholdnames}
\alias{createhallpwitholdnames}
\title{createhallpwitholdnames}
\usage{
createhallpwitholdnames()
}
\value{
Returns a collected copy of hallptbl as a dataframe.
}
\description{
Creates a copy of hallp with old column names, ie. prior to Fall 2015.
}

