% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchfunctions.R
\name{creatematchingcolumns}
\alias{creatematchingcolumns}
\title{Create matching strings.}
\usage{
creatematchingcolumns(x)
}
\arguments{
\item{x}{a data frame processed by prepformatch() or prephallp()}
}
\value{
a data frame with concatenated name fields which will be used for matching.
}
\description{
Builds the columns to do pidm matching on. Requires a df built from prepformatch() or prephallp().
}

