% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbReconnect.R
\name{dbReconnect}
\alias{dbReconnect}
\title{dbReconnect}
\usage{
dbReconnect()
}
\value{
returns TRUE invisibly
}
\description{
reconnect to a database and reinitialize all the tables
}
\details{
basically just works with cdw connections at this point
}
