% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.as.text.R
\name{df.as.text}
\alias{df.as.text}
\title{Make a data frame into a delimited text Table}
\usage{
df.as.text(df, sep = "\\t")
}
\arguments{
\item{df}{a data frame to convert to a vector}

\item{sep}{a delimiter. Defaults to tab-seperated text.}
}
\value{
a character vector with one entry per row of the data frame. Each row is collasped to a single entry with variables separated by tabs. NA and Inf values are replaced with blanks.
}
\description{
collapse a data frame into a tab-seperated text table, suitable for a .tsv file. Particularly useful for stringing multiple data frames together into a single report.
}
