% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{excel}
\alias{excel}
\title{Excel}
\usage{
excel(file)
}
\arguments{
\item{file}{a character string describing the path of the file to open. If the file does not open with Excel by default in Windows, it will open in the default program.}
}
\value{
returns NULL if successfull.
}
\description{
open a file in Excel. This is convience wrapper around `shell.exec(normalizePath(file) )`.
}

