% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillna.R
\name{fillna}
\alias{fillna}
\title{FillNA}
\usage{
fillna(x, fill = 0)
}
\arguments{
\item{x}{a vector with NA values to be filled}

\item{fill}{a value to replace any NA values in x. Note that if the class of fill does not match the class of x, x will be coerced to an acceptable class.}
}
\value{
a vector with length of x
}
\description{
Replaces NA values in a vector with a value, 0 by default. This is really just a convinence wrapper around replace() to minimize typing.
}
