% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchfunctions.R
\name{findlocation}
\alias{findlocation}
\title{Find columns}
\usage{
findlocation(df, matchtext)
}
\arguments{
\item{df}{a data frame that you want to find the location on}

\item{matchtext}{a regular expression to be matched against columns names}
}
\value{
returns a single integer value, the index of the first column that matches matchtext
}
\description{
Finds, within a data frame's columns, the first match for a series of regexp. Also send messages to the user to let them know what we matched on. Used within prepformatch().
}
\keyword{internal}

