% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchfunctions.R
\name{getpidmsoncatname}
\alias{getpidmsoncatname}
\title{Get pidm based on name.}
\usage{
getpidmsoncatname(x, y)
}
\arguments{
\item{x}{a data frame which needs pidms}

\item{y}{a data frame of reference data, usually hallp}
}
\value{
a data frame with nrows(x) with a new column, pidm_catname, appended
}
\description{
Returns a data frame with pidms based on catname appended.
}
