% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchfunctions.R
\name{getpidmsonemail}
\alias{getpidmsonemail}
\title{Get pidm based on email.}
\usage{
getpidmsonemail(x, y)
}
\arguments{
\item{x}{a data frame which needs pidms}

\item{y}{a data frame of reference data, usually hallp}
}
\value{
a data frame with nrows(x) with a new column, pidm_email, appended
}
\description{
Returns a data frame with pidms based on email appended based on matching rules.
}
