% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchfunctions.R
\name{getpidmsonlastaddr}
\alias{getpidmsonlastaddr}
\title{Get pidm baesd on lastname, address.}
\usage{
getpidmsonlastaddr(x, y)
}
\arguments{
\item{x}{a data frame which needs pidms}

\item{y}{a data frame of reference data, usually hallp}
}
\value{
a data frame with nrows(x) with a new column, pidm_lastaddr, appended
}
\description{
Returns a data frame with pidms based on last nameand address appended.
}

