% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matchfunctions.R
\name{getpidmsonlastfirst}
\alias{getpidmsonlastfirst}
\title{Get pidm based on lastname.}
\usage{
getpidmsonlastfirst(x, y)
}
\arguments{
\item{x}{a data frame which needs pidms}

\item{y}{a data frame of reference data, usually hallp}
}
\value{
a data frame with nrows(x) with a new column, pidm_lastfirst, appended
}
\description{
Returns a data frame with pidms based on lastname, first initial appended.
}

