% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getprimaries.R
\name{getprimaries}
\alias{getprimaries}
\title{Get Primary People}
\usage{
getprimaries(df, y = hallptbl, by = "pidm")
}
\arguments{
\item{df}{a data frame}

\item{y}{a hallp-like data frame or database table}

\item{by}{a character vector of variables to join by}
}
\value{
a one columns data frame of only primary people's pidms
}
\description{
turn a data frame of people into a data frame of only primary. Grabs any primary people that are on the list, plus the primary spouse if they're not already on the list.
}
