% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dupstatus.R
\name{getrelationships}
\alias{getrelationships}
\title{GetRelationships}
\usage{
getrelationships(df, idcol = "pidm", xrefslong = NA)
}
\arguments{
\item{df}{a data frame to work on with one line per constituent with one pidm column.}

\item{idcol}{a string naming the pidm column}

\item{xrefslong}{an object reference to a data frame built by buildlongxref() on the same data. If NA, this is generated from data/xaxref-pa.csv and data/xaxref-si.csv.}
}
\value{
a data frame with one row per person who has any relationship to someone else in the database. Constituents who are not related to anyone else are removed.
}
\description{
builds a table showing what relationships each individual in a data frame has with others in the same data frame. Used by dupstatus().
}

