% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmutheme.R
\name{ggmutheme}
\alias{ggmutheme}
\title{Set a Millikin style theme for ggplots}
\usage{
ggmutheme()
}
\description{
Creates a ggthemr theme based on the Millikin University pallette and applies it to all ggplots.
}

