% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giftsandmemos.R
\name{giftsandmemos}
\alias{giftsandmemos}
\title{Create a List of Gifts and memos}
\usage{
giftsandmemos(collectdata = F)
}
\arguments{
\item{collectdata}{a logical value describing whether or not the data should be collected from the database}
}
\value{
if collectdata == T, a data frame of all gifts and memos with one line per transaction. If collectdata is FALSE, a tbl_sql with 6 columns:
    - pidm - constituent's id
    - desg - the designation number to which the gift was given
    - campaign - the campaign in which the gift was given
    - fisc_code - the fiscal year in which the gift was given
    - amt - the gift amount or the amount which was memo credited to the constituent
    - dt - the gift date
}
\description{
Create a List of Gifts and memos
}

