% Generated by roxygen2 (4.0.2): do not edit by hand
\name{grepnames}
\alias{grepnames}
\title{Search for a string in column names}
\usage{
grepnames(str = ".*", x, ignorecase = T, sortalpha = F)
}
\arguments{
\item{str}{a regex string to find. Defaults to ".*"}

\item{x}{the object to search. Expect a data frame}

\item{ignorecase}{should the search ignore case. Default = T}
}
\description{
Performs a grep regex search on the column names of a data frame, returning the values and their column number.
}
\examples{
grepnames("lifetimeg",hallp)
grepnames("event",hallp,ignorecase=F)
}

