% Generated by roxygen2 (4.0.2): do not edit by hand
\name{histrug}
\alias{histrug}
\title{Plot a histogram with a rug}
\usage{
histrug(x)
}
\arguments{
\item{x}{a numeric vector to plot}
}
\description{
a convience wrapper to put a rug underneath a histogram. Also calls muplot for formatting tweaks
}

