% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initcommitsdb.R
\name{initcommitsdb}
\alias{initcommitsdb}
\title{Initialize my db connection}
\usage{
initcommitsdb(
  use = "RMariaDB",
  odbc = "warehouse510c",
  mysqldb = "commits",
  mysqlhost = "10.40.9.145",
  mysqluser = "adc",
  mysqlpassword = "goBigBlue",
  ...
)
}
\arguments{
\item{use}{what kind of database connection to use. Currently valid options include: \code{RMariaDB}, \code{odbc}, \code{mysql} and \code{pool}. RMariaDB is used by default. The pool option uses RMariaDB, but makes the connection with \code{pool::dbPool()} rather than \code{DBI::dbConnect()}.}

\item{odbc}{a string of an odbc connection to use. If NULL, database connects with credentials in mysql* arguments.}

\item{mysqldb}{a string of the name of the database name with which you should connect. Defaults to the 'commits' db.}

\item{mysqlhost}{the hostname or ip address for the database server}

\item{mysqluser}{the database username}

\item{mysqlpassword}{the database password}

\item{...}{further arguments to pass to DBI::dbConnect}
}
\value{
Assigns a database connection, plus connections to the hallp, dnrc, desgs, commits, gifts, pledges and memos tables in the global environment.
}
\description{
a convinence function to connect to the commits database on localhost.
}
