% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert.vector.R
\name{invert.vector}
\alias{invert.vector}
\title{invert.vector}
\usage{
invert.vector(x)
}
\arguments{
\item{x}{a numeric vector to be inverted}
}
\value{
a numeric vector with length = length(x)
}
\description{
Inverts a linear vector x, setting min(x) = min(x) and vice versa. Values in between are assumed to be linear values along the scale between min(x) and max(x).
}

