% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest.file.R
\name{latest.file}
\alias{latest.file}
\title{Identify the Most Recently Edited File}
\usage{
latest.file(path = ".", pattern = NULL, invert = F, ...)
}
\arguments{
\item{path}{a character vector of full path names; the default corresponds to the working directory, `getwd()`. Tilde expansion (see `path.expand``) is performed. Missing values will be ignored.}

\item{pattern}{an optional regular expression. Only file names which match the regular expression will be returned.}

\item{invert}{a logical value. If FALSE, returns the most recently modified file. If TRUE, returns the first modified file.}

\item{...}{other arguments to pass to `list.files`}
}
\value{
a character string with the full name of the most recently modified file in path
}
\description{
Identify the Most Recently Edited File
}

