% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchfunctions.R
\name{matchforpidms}
\alias{matchforpidms}
\title{Get new pidms}
\usage{
matchforpidms(x)
}
\arguments{
\item{x}{a data frame with most (preferably all) the columns to match on}
}
\value{
a data frame with nrows(x) with multiple columns appended, namely finalpidm, which has the function's best guess of what the pidms should be.
}
\description{
Uses several matching functions to append new pidms to an existing data frame. Input should be a data frame with most (preferably all) of the columns you'd need to match on. Note, this is a time consuming function, as it pulls and maninpulates all constituent data. d
}
\details{
Matching is performed on the following fields, in this order:
\enumerate{
\item existing pidm column
\item email
\item lastname, firstname
\item lastname, first initial
\item lastname, address (only the first 70\% of the adddress is used)
}
In all cases, matches are only made when one match is possible. Also, all text fields are set to lowercase and punctuation is removed. Note that missing data currently throws an error--if any columns are missing, we recommend adding a column with all NA values.
}
