% Generated by roxygen2 (4.0.2): do not edit by hand
\name{muplot}
\alias{muplot}
\title{Set plot options to Millikin defaults}
\usage{
muplot()
}
\description{
Sets a variety of plot options so plots come out looking consistent with Millikin University branding. Sets options via par().
}
\keyword{par}
\keyword{plot}

