% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_dups.R
\name{n_dups}
\alias{n_dups}
\title{Count Duplicates}
\usage{
n_dups(x)
}
\arguments{
\item{x}{a vector to count duplicates}
}
\value{
a integer number of duplicate values found in x
}
\description{
Count Duplicates
}
