% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/outliers.R
\name{outliers}
\alias{outliers}
\title{Determine the upper and lower outliers}
\usage{
outliers(x)
}
\arguments{
\item{x}{a numeric vector to be analysed.}
}
\value{
a list with two values:
\enumerate{
 \item bottom - the lower bound
 \item top - the upper bound
 }
}
\description{
Given a numeric vector, return the upper and lower bounds for outliers. Assumes IQR * 1.5 as bounds.
}

