% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchfunctions.R
\name{prepformatch}
\alias{prepformatch}
\title{Prep a data frame for matching.}
\usage{
prepformatch(x, pidm = c("pidm", "banner ID", "bannerID"),
  firstname = c("firstname", "first name", "^first$"),
  lastname = c("lastname", "last name", "^last$"),
  addr = c("address line 1", "address", "addr", "address1", "address 1",
  "addr1"), city = c("city", "town", "hometown"),
  st = c("state/province", "state", "^st$"), zip = c("zip code",
  "zipcode", "zip"), email = c("preferred email", "preferred e-mail",
  "pref email", "pref e-mail", "e-mail", "email"))
}
\arguments{
\item{x}{a data frame to be prepared}

\item{pidm}{a vector of regular expressions to try to match for pidm location}

\item{firstname}{a vector of regular expressions to try to match for firstname location}

\item{lastname}{a vector of regular expressions to try to match for lastname location}

\item{addr}{a vector of regular expressions to try to match for address location}

\item{city}{a vector of regular expressions to try to match for city location}

\item{st}{a vector of regular expressions to try to match for state location}

\item{zip}{a vector of regular expressions to try to match for zip code location}

\item{email}{a vector of regular expressions to try to match for email location}
}
\value{
a data frame with nrows(x) with the columns matched by the regular expressions.
}
\description{
Preps a data frame for matching by just grabbing the necessary columns. Currently does not handle missing columns/no matches very well. Needs to be improved.
}
