% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchfunctions.R
\name{prephallp}
\alias{prephallp}
\title{Pull hallp}
\usage{
prephallp(collectdata = T)
}
\arguments{
\item{collectdata}{boolean value determing whether or not hallp data should be collected from the database. Defaults to TRUE.}
}
\value{
a data frame of hallp data. Note that data is only collected, not prepped in any way.
}
\description{
Pulls hallp and gets it ready for matching.
}

