% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.dir.R
\name{read.dir}
\alias{read.dir}
\title{Read Files in a Directory}
\usage{
read.dir(path = ".", pattern = NULL, sep = ",", ...)
}
\arguments{
\item{path}{a character vector of full path names}

\item{pattern}{an optional regular expression. Only file names which match the regular expression will be returned. The names of the list items will have pattern removed from them}

\item{sep}{the field separator character for data frame-like files. Defaults to comma separated values.}

\item{...}{other arguments to pass to list.files()}
}
\value{
a list with one object per file. Files are read via read.tidy(). As such, files that are not data frame-ish may not be read correctly.
}
\description{
Read all the files in a directory into a list object.
}
