% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeoutput.R
\name{seedata}
\alias{seedata}
\title{See Output Directory}
\usage{
seedata(d = "data")
}
\value{
returns a NULL value when successfull, primarily useful for its side effect.
}
\description{
Opens a Windows directory to the path given to the dir argument, by default, the output/ directory. If the given directory does not exist, a warning is thrown and the current working directory is opened.
}
