% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeoutput.R
\name{seeoutput}
\alias{seeoutput}
\title{See Outpout Directory}
\usage{
seeoutput(dir = "output")
}
\arguments{
\item{dir}{the path to a directory to open}
}
\description{
Opens a Windows directory to the path given to the dir argument, by default, the output/ directory. If the given directory does not exist, a warning is thrown and the current working directory is opened.
}

