% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skel.R
\name{skel}
\alias{skel}
\title{Create Rproject Skeleton}
\usage{
skel(createbasescript = F, basescriptname = "00-source-all-files.R")
}
\arguments{
\item{createbasescript}{a logical value indicating whether or not a base script should be created. The base script will source all R scripts in the working directory}

\item{basescriptname}{a character string describing the file name of the base script. Defaults to '00-source-all-files.R'}
}
\value{
Does not return a value, but has side effects, namely creating directories and the base script.
}
\description{
Creates the folder structure and optional basescript for an R project. Creates two directories, data and output, assuming they don't already exist as well as the base script.
}

