% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcebeep.R
\name{sourcebeep}
\alias{sourcebeep}
\title{Source with Beep}
\usage{
sourcebeep(file, echo = T, ..., beep = 1)
}
\arguments{
\item{file}{a file to source}

\item{echo}{should the call to source(file) be made with the echo argument set to TRUE}

\item{...}{additional arguments to pass to \code{source()}}

\item{beep}{a numerical value to pass to \code{beepr::beep()}, indicating what what sound should indicate the process is finished}
}
\description{
Source an R file and beep when you're finished
}
