% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/startup.R
\name{startup}
\alias{startup}
\title{startup}
\usage{
startup(load = F, xlsx = F, MySQL = F)
}
\arguments{
\item{load}{a boolean value determining whether .RData ought to be loaded}

\item{xlsx}{a boolean value determining whether the xlsx package ought to be loaded}

\item{MySQL}{a boolean value determing whether the MySQL package ought to be loaded}
}
\description{
do the things I do at startup all the time
}

