% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_detect2.R
\name{str_detect2}
\alias{str_detect2}
\title{Str_detect with NA handling}
\usage{
str_detect2(string, pattern, default = F)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something coercible to one.}

\item{pattern}{Pattern to look for. See \code{\link[stringr]{str_detect}} .}

\item{default}{a default value to be used if string is NA}
}
\value{
a logical vector, assuming default is set to logical. Using a string or numeric value will coerce the value to the class of default.
}
\description{
Str_detect with NA handling
}
