% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-lists.R
\name{validate_list}
\alias{validate_list}
\title{Validate an Mail or Email List}
\usage{
validate_list(df, type = "mail", solicitation = T, h = hallptbl)
}
\arguments{
\item{df}{a data frame that is a mail or email list. Email lists should include at least pidm, and prefemail. Mail lists should include:
\enumerate{
\item pidm
\item addr1
\item st
\item zip
}}

\item{type}{a character string describing whether df is a "mail" or "email" list}

\item{solicitation}{a logical value describing whether or not df is a solicitation. Determines whether or not solicitation exclusion codes should be used.}

\item{h}{a database table or local data frame that looks like hallptbl}
}
\description{
Validate an Mail or Email List
}
