% Generated by roxygen2 (4.0.2): do not edit by hand
\name{view.sample}
\alias{view.sample}
\title{view.sample}
\usage{
view.sample(x, transpose = TRUE, sample.n = 3, sample.method = "head",
  stdout = FALSE)
}
\arguments{
\item{x}{an object to view. view.sample() is designed to work best with data frames but will work with vectors as well.}

\item{transpose}{a boolean value determining whether or not the data frame should be transposed. Defaults to TRUE so long data frames can be easily examined.}

\item{sample.n}{an integer value determining how many rows should be sampled}

\item{sample.method}{a string describing the way rows should be chosen from x. Valid values include:
- `head` uses head() to pull the first sample.n rows
- `sample` randomly chooses sample.n rows
- `tail` uses tail() to pull the last sample.n rows}

\item{output}{a boolean value determing whether the results should be sent to R's standard output. The default value, FALSE, sends the output to RStudio's `View()` function.}
}
\value{
unless stdout==TRUE, does not return a value. Otherwise, returns the sampled data frame
}
\description{
get a quick view of gigantic data.
}

