% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/write.clip.R
\name{write.clip}
\alias{write.clip}
\title{Write data to the clipboard}
\usage{
write.clip(x, sep = "\\t", na = "", ...)
}
\arguments{
\item{x}{an object to write}

\item{na}{the string to replace NA values. Defaults to "".}

\item{...}{other arguments to pass to write.table()}

\item{sepr}{a string to seperate columns, passed to write.table as sep=sepr. Defaults to \t, ie. tab-seperated columns.}
}
\description{
A convience wrapper for write.table that writes tab-seperated data to the Windows clipboard. Does not work in *nix.
}

