% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write.tidy.R
\name{write.tidy}
\alias{write.tidy}
\title{write.tidy}
\usage{
write.tidy(x, file = "", row.names = FALSE, na = "", ...)
}
\arguments{
\item{x}{the object to be written, preferably a matrix or data frame. If not, it is attempted to coerce x to a data frame.}

\item{file}{either a character string naming a file or a connection open for writing. "" indicates output to the console.}

\item{row.names}{either a logical value indicating whether the row names of x are to be written along with x, or a character vector of row names to be written.}

\item{na}{the string to use for missing values in the data.}

\item{...}{other values to be passed to write.csv}
}
\description{
writes a tidy csv file. Primarily a convience wrapper around write.csv, defaults to not writing rownames and NA = ""
}

