% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.to.ouput.R
\name{write.to.ouput}
\alias{write.to.ouput}
\title{Write To Output Directory}
\usage{
write.to.ouput(df, ...)
}
\arguments{
\item{df}{a data frame to write to disk}

\item{...}{additional arguments passed to \code{write.csv()}}
}
\value{

}
\description{
Write To Output Directory
}
\details{
The data frame will be written to the output directory with a filename of "output/the-object-name.R.YYYY-MM-DD.csv" with today's date.
}
