% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yesno.as.logical.R
\name{yesno.as.logical}
\alias{yesno.as.logical}
\title{yesno.as.logical}
\usage{
yesno.as.logical(v)
}
\arguments{
\item{v}{a character vector to convert}
}
\value{
a logical vector. Values of "Y" and "Yes" are converted to TRUE. All other values, including NA, are converted to FALSE.
}
\description{
Convert a vector of Yes/No or Yes/NA values to logical.
}
\examples{

foo  <- c("Y", NA, "Y", NA)
yesno.as.logical(foo)

foo  <- c("Y", NA, "Y", "Yes",NA, "No", "N","something else")
yesno.as.logical(foo)
}
