% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.df.R
\name{clean.df}
\alias{clean.df}
\title{Clean the names of a data frame}
\usage{
clean.df(df, keep.underscore = TRUE,
  pidmpattern = "bannerid|banner_id")
}
\arguments{
\item{df}{a data frame}

\item{keep.underscore}{a logical value indicating whether underscores be kept or removed}

\item{pidmpattern}{a regular expression to be matched against the names of df. Set to \code{NULL} to turn off pattern mathing.}
}
\value{
a data frame of class tbl_df.
}
\description{
Clean up a data frame by changing all column names to lower case, removing spaces and punctuation.
}
\details{
If no column named \code{pidm} is found and pidmpattern is not \code{NULL}, then we look if there's only
}
