% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coalesce}
\alias{coalesce}
\title{Return non-null values}
\usage{
coalesce(...)
}
\arguments{
\item{...}{objects to be used}
}
\value{
returns an object with the length of the first argument. Elements are not recycled.
}
\description{
a function, like the one in SQL, which takes arguments and returns the first non-NA value.
}

