% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce.R
\name{coalesce}
\alias{coalesce}
\title{Replace NA values}
\usage{
coalesce(x, ...)
}
\arguments{
\item{x}{The first value to coalesce.}

\item{...}{Other values to coalesce.}
}
\value{
A vector of the same length as {x}.
}
\description{
Returns the first non-NA argument, similar to the SQL function 
  coalesce. If a vector or matrix is passed as first argument,
  the remaining arguments are recycled to generate a vector/matrix of
  the same dimension, and coalescing is done element by element.
  Taken from krlmlr's [misc package](https://github.com/krlmlr/kimisc), referenced at [StackOverflow](http://stackoverflow.com/questions/19253820/how-to-implement-coalesce-efficiently-in-r).
}

