% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histrug.R
\name{histrug}
\alias{histrug}
\title{Plot a histogram with a rug}
\usage{
histrug(df, col)
}
\arguments{
\item{df}{either a data frame containing the column to be plotted, or a numeric vector to plot}

\item{col}{if df is not null, then the unquoted name of a column in df to plot}
}
\description{
a convience wrapper to put a rug underneath a histogram. Also calls muplot for formatting tweaks
}
