% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/initcommitsdb.R
\name{initcommitsdb}
\alias{initcommitsdb}
\title{Initialize my db connection}
\usage{
initcommitsdb()
}
\value{
Assigns a database connection, plus connections to the hallp commits, gifts, pledges and memos tables in the global environment.
}
\description{
a convinence function to connect to the commits database on localhost.
}

