% Generated by roxygen2 (4.0.2): do not edit by hand
\name{initcommitsdb}
\alias{initcommitsdb}
\title{Initialize my db connection}
\usage{
initcommitsdb()
}
\value{
Assigns a database connection, plus connections to the hallp commits, gifts, pledges and memos tables in the global environment.
}
\description{
a convinence function to connect to the commits database on localhost.
}

