% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initcommitsdb.R
\name{initcommitsdb}
\alias{initcommitsdb}
\title{Initialize my db connection}
\usage{
initcommitsdb(db = "commits", host = "10.40.9.144", user = "adc",
  password = "goBigBlue", ...)
}
\arguments{
\item{db}{a string of the name of the database name with which you should connect. Defaults to the 'commits' db.}

\item{host}{the hostname or ip address for the database server}

\item{user}{the database username}

\item{password}{the database password}

\item{...}{further arguments to pass to dplyr::src_mysql}
}
\value{
Assigns a database connection, plus connections to the hallp commits, gifts, pledges and memos tables in the global environment.
}
\description{
a convinence function to connect to the commits database on localhost.
}

