% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initcommitsdb.R
\name{initcommitsdb}
\alias{initcommitsdb}
\title{Initialize my db connection}
\usage{
initcommitsdb(db = "commits", ...)
}
\arguments{
\item{db}{a string of the name of the database name with which you should connect. Defaults to the 'commits' db.}

\item{...}{further arguments to pass to dplyr::src_mysql}
}
\value{
Assigns a database connection, plus connections to the hallp commits, gifts, pledges and memos tables in the global environment.
}
\description{
a convinence function to connect to the commits database on localhost.
}

