% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.hallp.R
\name{read.hallp}
\alias{read.hallp}
\title{Read hallp.csv}
\usage{
read.hallp(ifexists = T, existsname = "hallp")
}
\arguments{
\item{ifexists}{should the function check if the object exists. Default = F}

\item{existsname}{what object should the function look for, assuming ifexists = T}

\item{x}{the filename to load. Should be a csv}
}
\description{
A convience function around read.tidy specifically to read in G:/ALUMNI/Jake T/datawarehouse/data/hallp.csv.
}

