% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.tidy.R
\name{read.tidy}
\alias{read.tidy}
\title{Read csvs into tidy format}
\usage{
read.tidy(x = "clipboard", sep = "\\t", ifexists = F, existsname, ...)
}
\arguments{
\item{x}{the filename to load. Should be a csv. Defaults to reading from the clipboard.}

\item{sep}{the field separator character. If x == 'clipboard', sep defaults '\t'--otherwise, it defaults to ',' to read csv data.}

\item{ifexists}{should the function check if the object exists. Default = F}

\item{existsname}{what object should the function look for, assuming ifexists = T. If ifexists == T and existname is not empty, the function will return the object with name = existsname.}

\item{...}{additional parameters to pass to read.csv}
}
\value{
Returns a data frame will lower cased names and all punctuation removed. If dplyr is already loaded, it returns a data frame as a tbl_df--otherwise, it returns a simple data frame.
}
\description{
This is a convience wrapper for read.csv which lowercases column names and removes punctuation. It also has an option to only read the csv if the object does not exist.
}

