% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write.clip}
\alias{write.clip}
\title{Write data to the clipboard}
\usage{
write.clip(x, sepr = "\\t")
}
\arguments{
\item{x}{an object to write}

\item{sepr}{a string to seperate columns, passed to write.table as sep=sepr}
}
\description{
A convience wrapper for write.table that writes tab-seperated data to the clipboard. Does not work in *nix.
}

