% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.clip.R
\name{write.clip}
\alias{write.clip}
\title{Write data to the clipboard}
\usage{
write.clip(x, size = "4096", sep = "\\t", na = "", row.names = F,
  excel = F, ...)
}
\arguments{
\item{x}{an object to write}

\item{size}{the file size to use for the clipboard. Defaults to 4096 which is reasonably large.}

\item{na}{the string to replace NA values. Defaults to "".}

\item{row.names}{a logical value determining whether  row names be written to the file. Defaults to FALSE.}

\item{excel}{a logical value determing whether Excel should be opened immediately after copying the data to the clipboard.}

\item{...}{other arguments to pass to write.table()}

\item{sepr}{a string to seperate columns, passed to write.table as sep=sepr. Defaults to tab-seperated columns.}
}
\description{
A convience wrapper for write.table that writes tab-seperated data to the Windows clipboard. Does not work in *nix.
}

