% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.tidy.R
\name{write.tidy}
\alias{write.tidy}
\title{write.tidy}
\usage{
write.tidy(x, file = "", row.names = FALSE, na = "", excel = F,
  ...)
}
\arguments{
\item{x}{the object to be written, preferably a matrix or data frame. If not, it is attempted to coerce x to a data frame.}

\item{file}{either a character string naming a file or a connection open for writing. "" indicates output to the console.}

\item{row.names}{either a logical value indicating whether the row names of x are to be written along with x, or a character vector of row names to be written.}

\item{na}{the string to use for missing values in the data.}

\item{excel}{a logical value indicating whether the newly written file should be opened in Excel}

\item{...}{Arguments passed on to \code{utils::write.table}
\describe{
  \item{append}{logical. Only relevant if \code{file} is a character
    string.  If \code{TRUE}, the output is appended to the
    file.  If \code{FALSE}, any existing file of the name is destroyed.}
  \item{quote}{a logical value (\code{TRUE} or \code{FALSE}) or a
    numeric vector.  If \code{TRUE}, any character or factor columns
    will be surrounded by double quotes.  If a numeric vector, its
    elements are taken as the indices of columns to quote.  In both
    cases, row and column names are quoted if they are written.  If
    \code{FALSE}, nothing is quoted.}
  \item{eol}{the character(s) to print at the end of each line (row).
    For example, \code{eol = "\r\n"} will produce Windows' line endings on
    a Unix-alike OS, and \code{eol = "\r"} will produce files as expected by
    Excel:mac 2004.}
  \item{dec}{the string to use for decimal points in numeric or complex
    columns: must be a single character.}
  \item{col.names}{either a logical value indicating whether the column
    names of \code{x} are to be written along with \code{x}, or a
    character vector of column names to be written.  See the section on
    \sQuote{CSV files} for the meaning of \code{col.names = NA}.}
  \item{qmethod}{a character string specifying how to deal with embedded
    double quote characters when quoting strings.  Must be one of
    \code{"escape"} (default for \code{write.table}), in which case the
    quote character is escaped in C style by a backslash, or
    \code{"double"} (default for \code{write.csv} and
    \code{write.csv2}), in which case it is doubled.  You can specify
    just the initial letter.}
  \item{fileEncoding}{character string: if non-empty declares the
    encoding to be used on a file (not a connection) so the character data can
    be re-encoded as they are written.  See \code{\link{file}}.}
}}
}
\description{
Writes a tidy csv file. Primarily a convience wrapper around write.csv, defaults to not writing rownames and NA = "".
}
